Enable static type checking with Pyrefly (#2136)
Enables static type checking of torchtitan with
[pyrefly](https://github.com/facebook/pyrefly). Type checking the code
helps catch bugs earlier in the development cycle.

* Adds pyrefly to CI, as part of the linting workflow.
* Addresses ~100 type errors that can be fixed via local code changes
and updates to type annotations, and silences the rest with `# pyrefly:
ignore` suppression comments. Note that
https://github.com/pytorch/torchtitan/commit/325efd946f1cbea85e503f9e684b8c879891fc1a
contains all of the non-comment changes.

+++ torchtitan/protocols/model.py
@@ -37,9 +37,7 @@ class BaseModelArgs:
         pass
 
     @abstractmethod
-    def get_nparams_and_flops(
-        self, model: nn.Module, seq_len: int
-    ) -> tuple[int, float]:
+    def get_nparams_and_flops(self, model: nn.Module, seq_len: int) -> tuple[int, int]:
         pass
